python: Extend SWIG gpgme_{sub}key with a __repr__ method.
* lang/python/gpgme.i: added a genericrepr macro
--

To look nicer in Python's REPL.

We define a generic __repr__ as a SWIG macro and use that to extend some
defined SWIG objects.

The alternative would have been to write a custom __repr__ function for
each class but that would need to be changed everytime the object's
structure changes. The bindings should be easy to maintain, I guess.
This comes at the expense that the reprs are now relatively long and
contain, for example, both keyid and fingerprint. But I think it's a
reasonable choice to make.