python: Default pubkey and seckey arguments of op_genkey to NULL.
* lang/python/gpgme.i: Copied signature of op_genkey from gpgme.h and
defaulted arguments to NULL
* lang/python/tests/t-genkey.py: Implemented a test for default
arguments
* lang/python/tests/Makefile.am: Added test to XTESTS
--
As it seems to be perfectly fine, expected even, to call op_genkey with
NULL for both parameters, defaulting to those values makes calling it
easier.

Note that in the test we use a "TempContext" to leave the existing
keyring alone and not potentially disturb other tests by creating new
secret keys.